refactor(web): move HeroSection to sections directory
Move HeroSection component from components to sections, update to new
Hero, and export via index.ts.
This improves project structure and maintainability.
