+++ apps/web/src/app/sections/index.ts
@@ -1,0 +1,1 @@
+export * from './Hero';
